+++ CHANGES.md
@@ -4,7 +4,6 @@
 
 # Next
 
-- Add beforeInput and afterInput events to Context and Module
 - Better Logging functions
 
 ## Done
@@ -12,6 +11,7 @@
 - Add isPropagated optional parameter to dispatch function
 - Refactor and simplify API
 - Remove onDispatch event from Module and log helpers
+- Add beforeInput and afterInput events to Context and Module
 
 # v1.3.1